refactoring construction of webpack config file
