+++ config/webpack/entry.js
@@ -1,0 +1,7 @@
+module.exports = ({ isPackage, packageConfig: { source }, buildFiles }) => {
+	if (isPackage) {
+		return source;
+	}
+
+	return buildFiles;
+};
