+++ config/webpack/resolve.js
@@ -1,0 +1,7 @@
+module.exports = () => {
+	return {
+		alias: {
+			'lodash-es': 'lodash',
+		},
+	};
+};
