+++ config/webpack/target.js
@@ -1,0 +1,3 @@
+module.exports = ({ defaultTargets }) => {
+	return `browserslist:${defaultTargets.join(', ')}`;
+};
